fix up backport merge
